Replace toggle group with toggle buttons
Use toggle buttons on top of the inhibitor action instead of a toggle
group; this allows us to define the UI purely in blueprint, without any
template callbacks.

It also gives us more freedom to style the buttons, so we can make them
larger and add more spacing, which is more touch friendly.